chore: splits out blocks and array into their own suites
